Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    int a,b,product;
+    float a,b,product;
     product=0;
     scanf("%d%d",&a&b);
     product=a*b;
